Refactor authentication code utility classes
